refactor: extract NotificationContent component and move mock data to Storybook
- Extract reusable NotificationContent component from notification page
- Clean up notification page to use only real API data
- Add comprehensive Storybook stories for testing infinite scroll with mock data
- Improve component separation and testability
